add -p option to only print filenames
usecase: i've deleted some junk from my download directory, and want
to propagate those changes to the backup.
with -p i can now use the backup directory as source, and it'll print
the list of files which have been deleted since.
the output is redirected into a file which i can then edit (remove
the files i wanna keep on the backup side) and then use its contents
to pass to `rm`.